+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `prune-remote-emojis` command.
 - Prune remote emojis in background.
+- Added `limits.media.emoji_size_limit` configuration parameter.
 
 ### Changed